finishing

+++ app/src/main/java/com/example/fgapp/view_model/ControllersVM.kt
@@ -3,8 +3,8 @@ package com.example.fgapp.view_model
 interface ControllersVM {
     var isConnected: Boolean
     var aileron: Float
-    var rudder: Float
-    var throttle: Float
+    var rudder: Int
+    var throttle: Int
     var elevator: Float
     fun disconnect()
